Move tooltester ExceptionHandler to utils package
