+++ tooltester/src/main/java/io/dockstore/tooltester/client/cli/Client.java
@@ -20,12 +20,12 @@ import static io.dockstore.common.S3ClientHelper.getToolId;
 import static io.dockstore.common.S3ClientHelper.getVersionName;
 import static io.dockstore.tooltester.client.cli.JCommanderUtility.out;
 import static io.dockstore.tooltester.client.cli.JCommanderUtility.printJCommanderHelp;
-import static io.dockstore.tooltester.helper.ExceptionHandler.COMMAND_ERROR;
-import static io.dockstore.tooltester.helper.ExceptionHandler.DEBUG;
-import static io.dockstore.tooltester.helper.ExceptionHandler.exceptionMessage;
 import static io.dockstore.tooltester.runWorkflow.WorkflowRunner.GSON;
 import static io.dockstore.tooltester.runWorkflow.WorkflowRunner.printLine;
 import static io.dockstore.tooltester.runWorkflow.WorkflowRunner.uploadRunInfo;
+import static io.dockstore.utils.ExceptionHandler.COMMAND_ERROR;
+import static io.dockstore.utils.ExceptionHandler.DEBUG;
+import static io.dockstore.utils.ExceptionHandler.exceptionMessage;
 
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.MissingCommandException;
