+++ tooltester/src/main/java/io/dockstore/tooltester/runWorkflow/WorkflowRunner.java
@@ -19,11 +19,11 @@ package io.dockstore.tooltester.runWorkflow;
 import static io.dockstore.common.S3ClientHelper.createFileName;
 import static io.dockstore.common.metrics.MetricsDataS3Client.generateKey;
 import static io.dockstore.tooltester.client.cli.JCommanderUtility.out;
-import static io.dockstore.tooltester.helper.ExceptionHandler.API_ERROR;
-import static io.dockstore.tooltester.helper.ExceptionHandler.COMMAND_ERROR;
-import static io.dockstore.tooltester.helper.ExceptionHandler.GENERIC_ERROR;
-import static io.dockstore.tooltester.helper.ExceptionHandler.errorMessage;
-import static io.dockstore.tooltester.helper.ExceptionHandler.exceptionMessage;
+import static io.dockstore.utils.ExceptionHandler.API_ERROR;
+import static io.dockstore.utils.ExceptionHandler.COMMAND_ERROR;
+import static io.dockstore.utils.ExceptionHandler.GENERIC_ERROR;
+import static io.dockstore.utils.ExceptionHandler.errorMessage;
+import static io.dockstore.utils.ExceptionHandler.exceptionMessage;
 import static java.util.UUID.randomUUID;
 import static org.apache.commons.lang3.math.NumberUtils.max;
 import static org.apache.commons.lang3.math.NumberUtils.min;
